(Upstream) Fix some 64bit warnings
also: chmlib: arm64 and amd64 platforms support

(cherry picked from commit 4f2f0a3fe9e2c81edf0ee1854c6008397927d29d)
(cherry picked from commit 0c5d5921e860b6c9972e53bf7972cba428a1de3e)
(cherry picked from commit 4d2b7d5dff8e6c9e8e8598a5f04a38a75a2a3ab9)

+++ crengine/src/lvstring.cpp
@@ -2358,7 +2358,7 @@ lString8 & lString8::trim()
             (pchunk->buf8[lastns]==' ' || pchunk->buf8[lastns]=='\t');
             --lastns)
         ;
-    int newlen = lastns-firstns+1;
+    int newlen = (int)(lastns - firstns + 1);
     if (newlen == pchunk->len)
         return *this;
     if (pchunk->nref == 1)
@@ -2734,7 +2734,7 @@ int TrimDoubleSpaces(lChar16 * buf, int len,  bool allowStartSpace, bool allowEn
             state = 2;
         }
     }
-    return pdst - buf;
+    return (int)(pdst - buf);
 }
 
 lString16 & lString16::trimDoubleSpaces( bool allowStartSpace, bool allowEndSpace, bool removeEolHyphens )
@@ -3023,8 +3023,8 @@ void Utf8ToUnicode(const lUInt8 * src,  int &srclen, lChar16 * dst, int &dstlen)
             s++;
         }
     }
-    srclen = s - src;
-    dstlen = p - dst;
+    srclen = (int)(s - src);
+    dstlen = (int)(p - dst);
 }
 
 lString16 Utf8ToUnicode( const char * s ) {
@@ -4570,7 +4570,7 @@ bool lString8::startsWith( const char * substring ) const
 {
     if (!substring || !substring[0])
         return true;
-    int len = strlen(substring);
+    int len = (int)strlen(substring);
     if (length() < len)
         return false;
     const lChar8 * s1 = c_str();
@@ -4602,7 +4602,7 @@ bool lString8::endsWith( const lChar8 * substring ) const
 {
 	if ( !substring || !*substring )
 		return true;
-    int len = strlen(substring);
+    int len = (int)strlen(substring);
     if ( length() < len )
         return false;
     const lChar8 * s1 = c_str() + (length()-len);
